+++ src/Exception/NormalizationFailedException.php
@@ -3,7 +3,7 @@
 namespace Torr\SimpleNormalizer\Exception;
 
 /**
- * Generic failure exception, for usage inside of custom normalizers.
+ * Generic failure exception, for usage inside custom normalizers.
  */
 class NormalizationFailedException extends \RuntimeException implements NormalizerExceptionInterface
 {